+++ releasenotes.md
@@ -1,2 +1,3 @@
 Whats New:
 ----------------------
+Added formula status